Refactoring helpers.py
- Split helpers.py in modules with clear and single responsibilities
- Check for circular import issues
- Test importing logic
- Remove the giant module
